* gwork:
  Use m_viewRect.
  Fixing for(auto to be for (auto to match previous code
  Rename iterators in range-for loops
  More overrides, more range-for, some simple int -> size_t changes
  Note on OpenGL version.
  range based for-each
  const_iterator work
  Various fixes

+++ source/gwork/source/Anim.cpp
@@ -113,9 +113,9 @@ static float GetEased(float fTime, float fEase)
 {
     // Ease in and Out if ease is < 0
     if (fEase < 0)
-        return -fTime/2*(std::cosf(3.14159f*fTime)-1);
+        return -fTime/2*(std::cos(3.14159f*fTime)-1);
 
-    return std::powf(fTime, fEase);
+    return std::pow(fTime, fEase);
 }
 
 void Gwk::Anim::TimedAnimation::Think()
